@@ -14,44 +14,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^4.0.2",
-    "@angular/compiler": "^4.0.2",
-    "@angular/compiler-cli": "^4.0.2",
-    "@angular/core": "^4.0.2",
-    "@angular/forms": "^4.0.2",
-    "@angular/http": "^4.0.2",
-    "@angular/platform-browser": "^4.0.2",
-    "@angular/platform-browser-dynamic": "^4.0.2",
-    "@angular/router": "^4.0.2",
-    "@types/jasmine": "^2.5.47",
-    "@types/node": "^7.0.13",
-    "axios": "^0.16.1",
+    "@angular-devkit/build-angular": "^0.6.8",
+    "@angular/animations": "^6.0.5",
+    "@angular/cli": "^6.0.8",
+    "@angular/common": "^6.0.5",
+    "@angular/compiler": "^6.0.5",
+    "@angular/core": "^6.0.5",
+    "@angular/forms": "^6.0.5",
+    "@angular/http": "^6.0.5",
+    "@angular/platform-browser": "^6.0.5",
+    "@angular/platform-browser-dynamic": "^6.0.5",
+    "@angular/platform-server": "^6.0.5",
+    "@angular/router": "^6.0.5",
+    "@types/jasmine": "2.5.47",
+    "@types/node": "7.0.13",
+    "ajv": "^6.5.1",
+    "axios": "0.16.1",
     "bcrypt": "*",
     "body-parser": "*",
-    "codelyzer": "^2.1.1",
+    "codelyzer": "2.1.1",
     "core-js": "*",
     "cors": "*",
     "express": "*",
-    "jasmine-spec-reporter": "^4.0.0",
+    "jasmine-spec-reporter": "4.0.0",
     "jsonwebtoken": "*",
-    "karma": "^1.6.0",
-    "karma-remap-istanbul": "^0.6.0",
+    "karma": "1.6.0",
+    "karma-remap-istanbul": "0.6.0",
     "mongoose": "*",
     "morgan": "*",
-    "protractor": "^5.1.1",
+    "octavian": "2.2.0",
+    "protractor": "5.1.1",
     "rxjs": "*",
     "ts-helpers": "*",
-    "ts-node": "^3.0.2",
-    "tslint": "^5.1.0",
-    "typescript": "^2.3.0",
-    "zone.js": "^0.8.5"
+    "ts-node": "3.0.2",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "*",
-    "@angular/compiler-cli": "*",
-    "@types/jasmine": "*",
-    "@types/node": "*",
-    "codelyzer": "*",
+    "@angular/compiler-cli": "^6.0.5",
+    "@types/jasmine": "2.5.47",
+    "@types/node": "7.0.13",
+    "codelyzer": "2.1.1",
     "jasmine-core": "*",
     "jasmine-spec-reporter": "*",
     "karma": "*",
@@ -61,7 +63,7 @@
     "karma-remap-istanbul": "*",
     "protractor": "*",
     "ts-node": "*",
-    "tslint": "*",
-    "typescript": "*"
+    "tslint": "^4.5.1",
+    "typescript": "^2.7.2"
   }
 }
