@@ -1,52 +1,81 @@
-import { Injectable, OnInit }               from '@angular/core';
-import { Router }                           from '@angular/router';
-import { Headers, Response, Http }          from '@angular/http';
-import { Observable }                       from 'rxjs';
-import { User }                             from '../_models/user';
-import {BehaviorSubject}                    from 'rxjs/BehaviorSubject'
-import {environment}                        from '../../environments/environment';
 
-@Injectable()
+import { Observable, throwError }                                           from 'rxjs';
+import { catchError, map, tap }                                             from 'rxjs/operators';
+import { Injectable, OnInit }                                               from '@angular/core';
+import { Router }                                                           from '@angular/router';
+import { HttpClient, HttpErrorResponse }                                    from '@angular/common/http';
+import { User }                                                             from '../_models/user';
+import { environment }                                                      from '../../environments/environment';
+// import { NavbarComponent }                                               from '../navbar/navbar.component';
+
+@Injectable({
+  providedIn: 'root',
+})
 export class UserService implements OnInit {
   public user = new User;
-  public authenticated: boolean;
-  public adm: string;
-  public admin: boolean;
-  private headers = new Headers({ 'Content-Type': 'application/json' });
   private url = environment.apiUrlBase;
 
   constructor(
-    private http: Http,
+    private http: HttpClient,
     private router: Router
   ) { }
 
   ngOnInit() {
     this.checkLogin();
+    console.log("found user : ", localStorage.getItem('username') );
   }
 
   register(username: string, password: string, firstName: string, lastName: string, email: string): Observable<any> {
     console.log("make");
     return this.http
-      .post(this.url + "setup", JSON.stringify({ username: username, password: password, firstName: firstName, lastName: lastName, email: email }), { headers: this.headers })
-      .map(this.extractData)
-      .catch(this.handleError);
+      .post<any>(this.url + "setup", { username: username, password: password, firstName: firstName, lastName: lastName, email: email })
+      .pipe(
+        catchError(this.handleError)
+      );
+      //,{headers:this.headers});
+      // .post(this.url + "setup", JSON.stringify({ username: username, password: password, firstName: firstName, lastName: lastName, email: email }), { headers: this.headers })
+      // .map(this.extractData)
+      // .catch(this.handleError);    pre HttpClient way of doing this
   }
 
   login(username: string, password: string): Observable<any> {
     return this.http
-      .post(this.url + "auth", JSON.stringify({ username: username, password: password }), { headers: this.headers })
-      .map(this.extractData)
-      .catch(this.handleError);
+      .post<any>(this.url + "auth", { username: username, password: password})
+      .pipe(
+        catchError(this.handleError)
+      );
+  }
+
+  setAuth(data){
+    this.user.auth = data.success;
+    this.user.admin = data.admin;
+    this.user.token = data.token;
+    this.user.id = data.userId;
+    this.user.username = data.username;
+    localStorage.setItem('auth', data.success);
+    localStorage.setItem('token', data.token);
+    localStorage.setItem('username', data.username);
+    localStorage.setItem('userId', data.userId);
+    localStorage.setItem('admin', data.admin);
+  }
+
+  setUser(data){
+    console.log(data, ' : id sat' );
+    this.user = data.user;
+    this.user.auth = data.success;
   }
 
   checkUser(username: string): Observable<any> {
-    const token = localStorage.getItem('token')
-      ? '?token=' + localStorage.getItem('token') : '';
-    if (this.authenticated) {
+    var token = localStorage.getItem('token') ? '?token=' + localStorage.getItem('token') : '';
+    if (this.user.auth) {
       return this.http
-        .post(this.url + "user" + token, JSON.stringify({ username: username }), { headers: this.headers })
-        .map(this.extractData)
-        .catch(this.handleError);
+        .post<any>(this.url + "user" + token, {username: username})
+        .pipe(
+          catchError(this.handleError)
+        );
+        // .post(this.url + "user" + token, JSON.stringify({ username: username }), { headers: this.headers })
+        // .map(this.extractData)
+        // .catch(this.handleError); pre HttpClient way of doing this
     }
     else {
       alert("Please login");
@@ -56,12 +85,16 @@ export class UserService implements OnInit {
   }
 
   getUsers(): Observable<any> {
-    const token = localStorage.getItem('token')
+    var token = localStorage.getItem('token')
       ? '?token=' + localStorage.getItem('token') : '';
     return this.http
-      .get(this.url + "users" + token, { headers: this.headers })
-      .map(this.extractData)
-      .catch(this.handleError);
+      .get<any>(this.url + "users" + token)
+      .pipe(
+        catchError(this.handleError)
+      );;
+      // .get(this.url + "users" + token, { headers: this.headers })
+      // .map(this.extractData)
+      // .catch(this.handleError);
   }
 
   logout() {
@@ -71,34 +104,27 @@ export class UserService implements OnInit {
   }
 
   checkLogin() {
-    if (localStorage.getItem('token')) {
-      this.authenticated = true;
-      this.user.username = localStorage.getItem("username");
-      this.checkAdmin();
-    } else {
-      this.authenticated = false;
-    }
-  }
-  checkAdmin(){
-    this.adm = localStorage.getItem("admin");
+    this.user.auth = eval(localStorage.getItem('auth'));
+    this.user.token = localStorage.getItem('token');
+    this.user.username = localStorage.getItem('username');
+    this.user.id = localStorage.getItem('userId');
+    this.user.admin = eval(localStorage.getItem('admin'));  
   }
 
-  private extractData(res: Response) {
-    let body = res.json();
-    //console.log(body || {});
-    return body || {};
-  }
-
-  private handleError(error: Response | any) {
-    let errMsg: string;
-    if (error instanceof Response) {
-      const body = error.json() || '';
-      const err = body.error || JSON.stringify(body);
-      errMsg = `${error.status} - ${error.statusText || ''} ${err}`;
+  private handleError(error: HttpErrorResponse) {
+    if (error.error instanceof ErrorEvent) {
+      // A client-side or network error occurred. Handle it accordingly.
+      console.error('An error occurred:', error.error.message);
     } else {
-      errMsg = error.message ? error.message : error.toString();
+      // The backend returned an unsuccessful response code.
+      // The response body may contain clues as to what went wrong,
+      alert("There was an error, Please check the Console logs.")
+      console.error(
+        `Backend returned code ${error.status}, ` +
+        `body was: ${error.error}`);
     }
-    console.error(errMsg);
-    return Observable.throw(errMsg);
-  }
+    // return an observable with a user-facing error message
+    return throwError(
+      'Something bad happened; please try again later.');
+  };
 }
