@@ -2,7 +2,7 @@ import { Component, OnInit }      from '@angular/core';
 import { UserService }            from '../_services/user.service';
 import { Router }                 from '@angular/router';
 import { LoginComponent }         from '../login/login.component'
-import {BehaviorSubject}          from 'rxjs/BehaviorSubject'
+import {BehaviorSubject}          from 'rxjs'
 
 @Component({
   selector: 'navbar',
@@ -10,17 +10,17 @@ import {BehaviorSubject}          from 'rxjs/BehaviorSubject'
   styleUrls: ['./navbar.component.css']
 })
 export class NavbarComponent implements OnInit {
-  public isLoggedIn: boolean;
+
 
   constructor(private userService: UserService) { }
 
+ 
   ngOnInit() {
     this.userService.checkLogin();
   }
+  
   logout() {
-    this.userService.logout();
+    this.userService.logout(); 
   }
-  loggedIn(success) {
-    console.log(success + " logged in ");
-  }
+
 }
