@@ -12,6 +12,7 @@ export class ProfileComponent implements OnInit {
   errorMessage: string;
   username: string;
   token: string;
+  user = this.userService.user
 
   constructor(
     private userService: UserService,
@@ -19,19 +20,23 @@ export class ProfileComponent implements OnInit {
   ) { }
 
   ngOnInit() {
-    this.token = localStorage.getItem('token')
-    this.checkUser();
-  }
-  checkUser() {
-    this.username = localStorage.getItem('username');
-    if (!this.username) {
+    this.userService.checkLogin();
+    if(!this.userService.user){
       alert('need to login');
       this.router.navigateByUrl("/login");
     }
-    this.userService.checkUser(this.username)
+    this.getUser();
+    this.token = this.userService.user.token;
+  }
+  
+  getUser() {
+        this.userService.checkUser(this.userService.user.username)
       .subscribe(
       res => {
-        this.userService.user = res.user;
+        // this.userService.user = res.user;
+        console.log(res);
+        this.userService.setUser(res);
+        this.user = this.userService.user;
       },
       error => this.errorMessage = <any>error);
   }
