@@ -2,7 +2,8 @@ import { Component, OnInit }        from '@angular/core';
 import { Router }                   from '@angular/router';
 import { User }                     from '../_models/user';
 import { UserService }              from '../_services/user.service';
-import { Headers, Response, Http }  from '@angular/http';
+import { Headers, Response }        from '@angular/http';
+import { HttpClient }               from '@angular/common/http';
 
 @Component({
   selector: 'app-register',
@@ -34,7 +35,11 @@ export class RegisterComponent implements OnInit {
   }
 
   ngOnInit() {
-
+    this.userService.checkLogin();
+    if(this.userService.user.auth){
+      //already loggedin
+      this.router.navigateByUrl('/');
+    }
   }
 
 }
